seperate deal class
